script/python/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.0
pillow==10.3.0
proglog==0.1.10
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.2
